Cleanup code in FileSingleStreamSpiller
Move single-use utility method and cleanup exception handling
